apps/studiocentos/apps/backend/pyproject.toml
[project]
name = "studiocentos-backend"
version = "0.1.0"
description = "StudiocentOS Backend API - Portfolio, Booking & AI Services"
authors = [{name = "Ciro Autuori", email = "ciro@studiocentos.it"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core Framework
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    # Database & ORM
    "sqlalchemy>=2.0.35",
    "pydantic[email]>=2.10.4",
    "pydantic-settings>=2.11.0",
    "psycopg2-binary>=2.9.10",
    "alembic>=1.14.0",
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bleach>=6.2.0",
    "cryptography>=44.0.0",
    # Utilities
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.17",
    "pytz>=2024.2",
    "aiofiles>=24.1.0",
    # Background Tasks
    "apscheduler>=3.11.0",
    # Admin & Monitoring
    "sqladmin>=0.21.0",
    # File Storage & Reporting
    "fastapi-storages>=0.3.0",
    "reportlab>=4.3.1",
    "jinja2>=3.1.2",
    # Performance & Caching
    "redis>=5.2.1",
    # Skills Enhancement
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.26.0",
    # Monitoring & Observability
    "structlog>=24.4.0",
    "sentry-sdk[fastapi]>=2.0.0",
    # Payments & Auth
    "stripe>=13.0.1",
    "pyotp>=2.9.0",
    "qrcode>=8.2",
    "pillow>=12.0.0",
    "aiohttp>=3.13.0",
    # AI - SOLO API HTTP (GROQ, Google, OpenRouter, HuggingFace)
    "httpx>=0.28.1",
    "groq>=0.9.0",
    "google-generativeai>=0.8.3",
    # Google Client APIs (Calendar, Drive, etc.)
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.111.0",
    # ML Embeddings & Vector Store
    "chromadb>=0.4.22",
    "openai>=1.50.0",
    # Social Media APIs
    "tweepy>=4.14.0",
    "facebook-sdk>=3.1.0",
    # Web Scraping for Lead Generation
    "beautifulsoup4>=4.12.3",
    "lxml>=5.3.0",
    # ToolAI - Real data scraping
    "feedparser>=6.0.10",
]

[project.optional-dependencies]
dev = [
    # Code Quality & Linting
    "ruff>=0.8.5",
    "black>=24.10.0",
    "isort>=5.13.2",
    "mypy>=1.13.0",
    # Testing
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.1",
    # Development Utilities
    "pre-commit>=4.0.1",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*", "alembic*"]

# === CODE QUALITY CONFIGURATION ===

[tool.ruff]
select = ["E", "W", "F", "I", "N", "B", "A", "C", "D", "UP", "S", "T"]
ignore = ["D100", "D101", "D102", "D103", "D104", "D105"]
line-length = 100
target-version = "py312"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "D415"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
addopts = "-v --cov=app --cov-report=term-missing --cov-report=html --cov-report=xml --cov-fail-under=40"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests (fast, no external deps)",
    "integration: Integration tests (DB, external APIs)",
]

[tool.coverage.run]
branch = true
source = ["app"]
omit = [
    "*/tests/*",
    "*/alembic/*",
    "*/__init__.py",
    "*/migrations/*",
    "app/main.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "def __str__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", "migrations", ".venv", "venv"]
skips = [
    "B101",
    "B601",
]
confidence_level = "MEDIUM"
